#02: defining functions
Added the function to calculate the cost of the project. A constant
Calculator was created to contain the costOfProject function, which
will be exported to the test.js file to test. I set up basic parameters
for the function, such as checking to see if the base price/ input
price is valid (if its not, then throw an error. if its a string, then
parse it. if its not a number, throw an error). Made variables to
calculate the cost, and returned it to the function as a float with 2
decimal places. A function for calculating the markup per person and
categories was also made. The category type used a switch-case
function. I may change this to an object literal table to match with
current trends of javascript (if time permits).