#03: Set up test.js
Erased the test var assets (from mocha testing installation), imported
the Calculator constant from index.js, and made a variable to require
chai testing. Chai.should() is used to test for the parameters. Next
steps is to include the testing parameters. Some of the thing we can
test for include:
1) if the input price, workers required and category equal to the
expected outcome….we have 3 input variables from the coding challenge,
so x3.
2)we should also check to see that the base price/input price is valid
(aka a number)
3) check to see if each category gives you the correct markup
percentage. we have 3 categories and 1 default….so x4
4)we should check to see if the workers required percent markup is
correct, and set a fallback number of worker. Im not sure if we should
assume there are no workers if number of workers is not indicated, or
should we assume there’s always 1 worker needed???? May need to clarify
with nulogy.