#03: fixed errors on test.js
Fixed the last error in the test parameters. The test expected to throw
an error when the basePrice is not a number or is invalid, however, it
did not do so because I had 2 different throw errors for that situation
in my index.js file and in my test.js file. I removed the extra throw
error, and the test passed.