#04: Refactoring to ES6
Changed my code to refactor it into ES6. I made the constant CALCULATOR
into a class Calculator, and then exported it into the test.js file. In
the test.js file, I imported the class Calculator, and then made a new
instance of Calculator which I name Calc. The functions and test ran as
expected.